Fixed an issue where including the file's time in the hash could prevent some changes from being applied.
Specifically: if a file was hmr unacceptable and had changes applied they would be ignored. If those changes were then undone, any further changes to other files would not be applied because the unacceptable file now presented a different hash. This rolls from the assumption that all files must be acceptable to apply the changes.